Refactor: Update import paths to use capitalized directory names
