+++ App.tsx
@@ -6,21 +6,21 @@ import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { Toaster } from 'sonner-native';
 import { AuthProvider, useAuth } from './hooks/useAuth';
 import type { RootStackParamList } from './types/navigation';
-import WelcomeScreen from './screens/auth/WelcomeScreen';
-import LoginScreen from './screens/auth/LoginScreen';
-import SignupScreen from './screens/auth/SignupScreen';
-import MainTabs from './screens/MainTabs';
-import RunScreen from './screens/RunScreen';
-import UpgradeScreen from './screens/UpgradeScreen';
-import ConnectSpotifyScreen from './screens/ConnectSpotifyScreen';
-import ConnectWatchScreen from './screens/ConnectWatchScreen';
+import WelcomeScreen from './Screens/auth/WelcomeScreen';
+import LoginScreen from './Screens/auth/LoginScreen';
+import SignupScreen from './Screens/auth/SignupScreen';
+import MainTabs from './Screens/MainTabs';
+import RunScreen from './Screens/RunScreen';
+import UpgradeScreen from './Screens/UpgradeScreen';
+import ConnectSpotifyScreen from './Screens/ConnectSpotifyScreen';
+import ConnectWatchScreen from './Screens/ConnectWatchScreen';
 import { ThemeProvider, useTheme } from './hooks/useTheme';
 import { GateProvider } from './hooks/useGate';
-import { setApiBaseUrl } from './lib/api';
-import { API_BASE_URL } from './lib/config';
-import RunSummaryScreen from './screens/RunSummaryScreen';
-import RouteDetailScreen from './screens/routes/RouteDetailScreen';
-import SavedRoutesScreen from './screens/routes/SavedRoutesScreen';
+import { setApiBaseUrl } from './Lib/api';
+import { API_BASE_URL } from './Lib/config';
+import RunSummaryScreen from './Screens/RunSummaryScreen';
+import RouteDetailScreen from './Screens/routes/RouteDetailScreen';
+import SavedRoutesScreen from './Screens/routes/SavedRoutesScreen';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
