+++ components/RunList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet, Pressable } from 'react-native';
 import { useTheme } from '../hooks/useTheme';
-import type { Run } from '../lib/runStore';
+import type { Run } from '../Lib/runStore';
 
 let SvgLib: any = null;
 try { SvgLib = require('react-native-svg'); } catch {}
@@ -19,12 +19,12 @@ export default function RunList({ runs, onPressItem }: { runs: Run[]; onPressIte
           </View>
           <Text style={{ color: theme.colors.muted }}>{r.distanceKm.toFixed(2)} km • {r.avgPace} min/km • {Math.round(r.calories)} kcal</Text>
           {SvgLib && r.splits && r.splits.length > 1 && (() => {
-            const secs = r.splits!.map((s) => s.sec);
+            const secs = r.splits!.map((s) => s.paceSec);
             const minS = Math.min(...secs);
             const maxS = Math.max(...secs);
             const toStr = (r.splits!.map((s, i) => {
               const x = i / (r.splits!.length - 1);
-              const yn = (s.sec - minS) / ((maxS - minS) || 1);
+              const yn = (s.paceSec - minS) / ((maxS - minS) || 1);
               const y = 1 - yn; // invert so faster pace is higher
               return `${x},${y}`;
             }).join(' '));
